refactor/feat: изменить структуру пакетов/добавить BotManager и BtCommandStrategy
