+++ src/main/java/ru/paraktikumbot/bot/main/botmanager/service/BotManagerService.java
@@ -1,0 +1,13 @@
+package ru.paraktikumbot.bot.main.botmanager.service;
+
+import ru.paraktikumbot.bot.main.common.model.Update;
+
+public class BotManagerService {
+
+    public void process(Update update) {
+
+    }
+
+
+
+}
