+++ src/main/java/ru/paraktikumbot/bot/main/botmanager/strategy/BotCommandContext.java
@@ -1,0 +1,4 @@
+package ru.paraktikumbot.bot.main.botmanager.strategy;
+
+public class BotCommandContext {
+}
