+++ src/main/java/ru/paraktikumbot/bot/main/botmanager/strategy/HelloCommand.java
@@ -1,0 +1,16 @@
+package ru.paraktikumbot.bot.main.botmanager.strategy;
+
+import ru.paraktikumbot.bot.main.common.model.Update;
+
+public class HelloCommand implements BotCommandStrategy {
+    
+    @Override
+    public boolean apply(Update update) {
+        return false;
+    }
+
+    @Override
+    public void process(Update update) {
+
+    }
+}
